refactor(graph): Extract persistGraph to separate module
- Moved persistGraph logic from page.tsx to a new module
  at giselle-provider/graph/persist.ts for better modularity
- Replaced inline persistGraph logic with import from new module
- Simplified graph handling in page.tsx by outsourcing blob deletion
  and graph URL update functionality

This refactor should make the codebase more maintainable by adhering
to the single responsibility principle, without introducing breaking
changes.
